fixed axis < 0 exception

+++ matrix_operations.h
@@ -19,7 +19,7 @@ namespace wef {
         Tensor argmax(const Tensor& m1);
         Tensor activation(const Tensor& m1, const char ops);
         Tensor softmax(const Tensor& m1);
-        Tensor reducesum(const Tensor& m1, const size_t ax);
+        Tensor reducesum(const Tensor& m1, const int ax);
         float l2(const Tensor& m1, const Tensor& m2);
         float binarycrossentropy(const Tensor& m1, const Tensor& m2);
         float categoricalcrossentropy(const Tensor& m1, const Tensor& m2, Tensor& m);
